requirements.txt
aiohttp==3.2.1
aniso8601==3.0.0
asn1crypto==0.24.0
async-timeout==3.0.0
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cffi==1.11.5
chardet==3.0.4
cryptography==1.9
Django==2.0.5
django-crontab==0.7.1
ecdsa==0.13
future==0.16.0
gevent==1.2.2
graphene==2.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
greenlet==0.4.13
gunicorn==19.8.1
idna==2.6
idna-ssl==1.0.1
iso8601==0.1.12
lxml==4.2.1
multidict==4.3.1
promise==2.1
pyasn1==0.4.2
pycparser==2.18
python-jose==3.0.0
pytz==2018.4
rsa==3.4.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
urllib3==1.22
yarl==1.2.4
